requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
_python_abi3_support=1.0=hd8ed1ab_2
asttokens=3.0.0=pyhd8ed1ab_1
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.10.5=h4c7d964_0
certifi=2025.10.5=pypi_0
charset-normalizer=3.4.4=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.3=pyhe01879c_0
contourpy=1.3.3=pypi_0
cpython=3.13.9=py313hd8ed1ab_101
cycler=0.12.1=pypi_0
debugpy=1.8.17=py313h927ade5_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
expat=2.7.1=h8ddb27b_0
filelock=3.20.0=pypi_0
fonttools=4.60.1=pypi_0
fsspec=2025.9.0=pypi_0
idna=3.11=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=7.0.1=pyh6dadd2b_0
ipython=9.6.0=pyh6be1c34_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.9.1=pyh6dadd2b_0
kiwisolver=1.4.9=pypi_0
krb5=1.21.3=hdf4eb48_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
libsodium=1.0.20=hc70643c_0
libzlib=1.3.1=h02ab6af_0
markupsafe=3.0.3=pypi_0
matplotlib=3.10.7=pypi_0
matplotlib-inline=0.2.1=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.5=pypi_0
numpy=2.2.6=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.5.4=h725018a_0
packaging=25.0=pyh29332c3_1
parso=0.8.5=pyhcf101f3_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=12.0.0=pypi_0
pip=25.2=pyhc872135_1
platformdirs=4.5.0=pyhcf101f3_0
polars=1.34.0=pypi_0
polars-runtime-32=1.34.0=pypi_0
prompt-toolkit=3.0.52=pyha770c72_0
psutil=7.1.1=py313h5ea7bf4_0
pure_eval=0.2.3=pyhd8ed1ab_1
pygments=2.19.2=pyhd8ed1ab_0
pyparsing=3.2.5=pypi_0
python=3.13.9=h260b955_100_cp313
python-dateutil=2.9.0.post0=pyhe01879c_2
python-gil=3.13.9=h4df99d1_101
python_abi=3.13=1_cp313
pywin32=311=py313h40c08fc_1
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=py312hbb5da91_0
requests=2.32.5=pypi_0
scipy=1.16.2=pypi_0
setuptools=80.9.0=py313haa95532_0
six=1.17.0=pyhe01879c_1
sqlite=3.50.2=hda9a48d_1
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.14.0=pypi_0
tk=8.6.15=hf199647_0
torch=2.7.1+cu118=pypi_0
torchaudio=2.7.1+cu118=pypi_0
torchvision=0.22.1+cu118=pypi_0
tornado=6.5.2=py313h5ea7bf4_1
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
ultralytics=8.3.221=pypi_0
ultralytics-thop=2.0.17=pypi_0
urllib3=2.5.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
wcwidth=0.2.14=pyhd8ed1ab_0
wheel=0.45.1=py313haa95532_0
xz=5.6.4=h4754444_1
zeromq=4.3.5=h5bddc39_9
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.3.1=h02ab6af_0
